+++ package.json
@@ -17,7 +17,8 @@
     "react-scripts": "4.0.1",
     "typescript": "4.1.3",
     "web-vitals": "1.0.1",
-    "xstate": "4.15.1"
+    "xstate": "4.15.1",
+    "axios": "0.21.0"
   },
   "scripts": {
     "start": "react-scripts start",
